TASK: Move upload and replace features into subpackage
