+++ Resources/Private/JavaScript/asset-upload/src/components/FilePreview.tsx
