+++ Resources/Private/JavaScript/asset-upload/src/components/PreviewSection.tsx
