+++ Resources/Private/JavaScript/asset-upload/src/components/index.ts
@@ -1,0 +1,19 @@
+import NewAssetDialog from './Dialogs/NewAssetDialog';
+import ReplaceAssetDialog from './Dialogs/ReplaceAssetDialog';
+import UploadDialog from './Dialogs/UploadDialog';
+import FilePreview from './FilePreview';
+import PreviewSection from './PreviewSection';
+import UploadSection from './UploadSection';
+import AssetReplacementButton from './AssetReplacementButton';
+import UploadButton from './UploadButton';
+
+export {
+    AssetReplacementButton,
+    FilePreview,
+    NewAssetDialog,
+    PreviewSection,
+    ReplaceAssetDialog,
+    UploadButton,
+    UploadDialog,
+    UploadSection,
+};
