+++ Resources/Private/JavaScript/asset-upload/src/hooks/useReplaceAsset.ts
@@ -1,7 +1,9 @@
 import { useMutation } from '@apollo/client';
 
+import { Asset } from '@media-ui/core/src/interfaces';
+
 import { REPLACE_ASSET } from '../mutations';
-import { Asset, FileUploadResult } from '../interfaces';
+import { FileUploadResult } from '../interfaces';
 
 export interface AssetReplacementOptions {
     generateRedirects: boolean;
