+++ Resources/Private/JavaScript/asset-upload/src/hooks/useUploadFile.ts
