+++ Resources/Private/JavaScript/asset-upload/src/hooks/useUploadFiles.ts
