+++ Resources/Private/JavaScript/asset-upload/src/interfaces/FileUploadResult.ts
