+++ Resources/Private/JavaScript/asset-upload/src/interfaces/index.ts
@@ -1,0 +1,3 @@
+import FileUploadResult from './FileUploadResult';
+
+export { FileUploadResult };
