+++ Resources/Private/JavaScript/asset-upload/src/mutations/index.ts
@@ -1,0 +1,5 @@
+import REPLACE_ASSET from './replaceAsset';
+import UPLOAD_FILES from './uploadFiles';
+import UPLOAD_FILE from './uploadFile';
+
+export { UPLOAD_FILES, UPLOAD_FILE, REPLACE_ASSET };
