+++ Resources/Private/JavaScript/asset-upload/src/mutations/replaceAsset.ts
