+++ Resources/Private/JavaScript/asset-upload/src/mutations/uploadFile.ts
