+++ Resources/Private/JavaScript/asset-upload/src/mutations/uploadFiles.ts
