+++ Resources/Private/JavaScript/asset-upload/src/state/index.ts
@@ -1,0 +1,3 @@
+import uploadDialogVisibleState from './uploadDialogVisibleState';
+
+export { uploadDialogVisibleState };
