+++ Resources/Private/JavaScript/asset-upload/src/state/uploadDialogVisibleState.ts
