+++ Resources/Private/JavaScript/asset-upload/tsconfig.json
@@ -1,0 +1,8 @@
+{
+  "extends": "../../../../tsconfig.json",
+  "compilerOptions": {
+    "rootDir": "src",
+    "module": "esnext",
+    "jsx": "react"
+  }
+}
