+++ Resources/Private/JavaScript/core/src/components/index.ts
@@ -1,0 +1,3 @@
+import { Dialog } from './Dialog';
+
+export { Dialog };
