+++ Resources/Private/JavaScript/core/src/helper/humanFileSize.ts
