+++ Resources/Private/JavaScript/core/src/helper/index.ts
