+++ Resources/Private/JavaScript/core/src/hooks/index.ts
@@ -11,7 +11,6 @@ import useDeleteAssetCollection from './useDeleteAssetCollection';
 import useDeleteTag from './useDeleteTag';
 import useEvent from './useEvent';
 import useImportAsset from './useImportAsset';
-import useReplaceAsset from './useReplaceAsset';
 import useSelectAsset from './useSelectAsset';
 import useSelectAssetCollection from './useSelectAssetCollection';
 import useSelectAssetSource from './useSelectAssetSource';
@@ -26,8 +25,6 @@ import useUntagAsset from './useUntagAsset';
 import useUpdateAsset from './useUpdateAsset';
 import useUpdateAssetCollection from './useUpdateAssetCollection';
 import useUpdateTag from './useUpdateTag';
-import useUploadFile from './useUploadFile';
-import useUploadFiles from './useUploadFiles';
 
 export {
     useAssetCollectionsQuery,
@@ -43,7 +40,6 @@ export {
     useDeleteTag,
     useEvent,
     useImportAsset,
-    useReplaceAsset,
     useSelectAsset,
     useSelectAssetCollection,
     useSelectAssetSource,
@@ -58,6 +54,4 @@ export {
     useUpdateAsset,
     useUpdateAssetCollection,
     useUpdateTag,
-    useUploadFile,
-    useUploadFiles,
 };
