+++ Resources/Private/JavaScript/core/src/interfaces/index.ts
@@ -5,7 +5,6 @@ import AssetIdentity from './AssetIdentity';
 import AssetSource from './AssetSource';
 import DeleteAssetCollectionResult from './DeleteAssetCollectionResult';
 import FeatureFlags from './FeatureFlags';
-import FileUploadResult from './FileUploadResult';
 import GraphQlEntity from './GraphQLEntity';
 import Image from './Image';
 import IptcProperty from './IptcProperty';
@@ -20,7 +19,6 @@ export {
     AssetSource,
     DeleteAssetCollectionResult,
     FeatureFlags,
-    FileUploadResult,
     GraphQlEntity,
     Image,
     IptcProperty,
