+++ Resources/Private/JavaScript/media-module/package.json
@@ -17,6 +17,7 @@
     "@media-ui/feature-concurrent-editing": "*",
     "@media-ui/feature-similar-assets": "*",
     "@media-ui/feature-asset-preview": "*",
+    "@media-ui/feature-asset-upload": "*",
     "@neos-project/react-ui-components": "^5.3.4",
     "apollo-upload-client": "^14.1.3",
     "classnames": "^2.2.6",
