+++ Resources/Private/JavaScript/media-module/src/components/Dialogs/CreateAssetCollectionDialog.tsx
@@ -5,11 +5,10 @@ import { Button, TextInput, Label } from '@neos-project/react-ui-components';
 
 import { useIntl, createUseMediaUiStyles, useNotify } from '@media-ui/core/src';
 import { useCreateAssetCollection } from '@media-ui/core/src/hooks';
+import { Dialog } from '@media-ui/core/src/components';
 
 import { useCallback } from 'react';
 import { createAssetCollectionDialogState } from '../../state';
-
-import { Dialog } from './Dialog';
 
 const useStyles = createUseMediaUiStyles(() => ({
     formBody: {
