+++ Resources/Private/JavaScript/media-module/src/components/Dialogs/index.ts
@@ -1,5 +1,4 @@
 import CreateTagDialog from './CreateTagDialog';
-import UploadDialog from './UploadDialog';
 import CreateAssetCollectionDialog from './CreateAssetCollectionDialog';
 
-export { CreateTagDialog, UploadDialog, CreateAssetCollectionDialog };
+export { CreateTagDialog, CreateAssetCollectionDialog };
