+++ Resources/Private/JavaScript/media-module/src/components/SideBarLeft/index.ts
@@ -2,6 +2,5 @@ import AssetSourceList from './AssetSourceList';
 import AssetSourceDescription from './AssetSourceDescription';
 import SideBarLeft from './SideBarLeft';
 import { AssetCollectionTree } from './Tree';
-import UploadButton from './UploadButton';
 
-export { AssetSourceList, AssetSourceDescription, SideBarLeft, AssetCollectionTree, UploadButton };
+export { AssetSourceList, AssetSourceDescription, SideBarLeft, AssetCollectionTree };
