+++ Resources/Private/JavaScript/media-module/src/state/index.ts
@@ -1,12 +1,5 @@
 import createAssetCollectionDialogState from './createAssetCollectionDialogState';
 import createTagDialogState from './createTagDialogState';
 import { mainViewState, MainViewState } from './mainViewState';
-import uploadDialogVisibleState from './uploadDialogVisibleState';
 
-export {
-    createAssetCollectionDialogState,
-    createTagDialogState,
-    mainViewState,
-    MainViewState,
-    uploadDialogVisibleState,
-};
+export { createAssetCollectionDialogState, createTagDialogState, mainViewState, MainViewState };
